requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
psycopg2-binary==2.9.6
pytest==7.2.2
pytz==2023.3
sqlparse==0.4.3
